refactor: rename bots/cleanup.py to bots/cleanup_utils.py
Aligns with the repo's naming convention for utility modules.
Updates the four call sites and the test file's import.
